requirements.txt
langchain
langchain-openai
langchain-chroma
python-dotenv
pypdf
faiss-cpu
faiss-gpu